* lua compile fixes
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2372 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/AIInterface.cpp
@@ -2674,7 +2674,7 @@ SpellCastTargets AIInterface::setSpellTargets(SpellEntry *spellInfo, Unit* targe
 
 AI_Spell *AIInterface::getSpell()
 {
-	if(next_spell_time > UNIXTIME)
+	if(next_spell_time > (uint32)UNIXTIME)
 		return NULL;
 
 	waiting_for_cooldown = false;
